BREAKING CHANGE: refactor wakeup and node addressing

+++ linux/apps/device-nrf24l01/main.cpp
@@ -27,7 +27,7 @@ void parseOpt(int argc, char* argv[], monitor& mon, LinuxCryptoHandler& cryptoHa
 {
     char option = 0;
     uint8_t spiRegister = 0;
-    uint8_t radioAddress = 0;
+    uint32_t radioAddress = 0;
 
     while ((option = getopt(argc, argv, "K:N:ir:s:c:C:h")) != -1) {
         switch (option) {
