+++ linux/apps/test/main.cpp
@@ -93,7 +93,7 @@ void parseOpt(int argc, char* argv[], monitor& mon, LinuxCryptoHandler& cryptoHa
     char option = 0;
     uint16_t i2cDeviceOffset = 0;
     uint8_t i2cDeviceAddress = 0b10100000;
-    uint8_t radioAddress = 0;
+    uint32_t radioAddress = 0;
     bool verbose = false;
 
     while ((option = getopt(argc, argv, "K:N:vVACeI:gbSshT:t")) != -1) {
